requirements.txt
discord.py
yfinance
pandas
numpy
matplotlib
scipy
python-dotenv
moomoo-api